Compiler can now generate small AST_NODE code segments if asked. Updated typechecker to properly check more types. Fixed bug where type-checking field access was not entirely possible.

+++ AAModifier.h
@@ -37,7 +37,7 @@ enum class AAAccessModifier : unsigned char {
 /// <summary>
 /// The storage modifiers
 /// </summary>
-enum class AAStorageModifier : long long {
+enum class AAStorageModifier : int16_t {
 
 	/// <summary>
 	/// No modifier
